+++ samples/07.advanced-web-chat-apps/e.sso-on-behalf-of-authentication/rest-api/src/utils/createHTMLWithPostMessage.js
@@ -1,0 +1,4 @@
+module.exports = (data, origin) =>
+  `<!DOCTYPE html><html><head><title></title><script>opener.postMessage('${new URLSearchParams(
+    data
+  )}', '${origin}');close();</script></head><body></body></html>`;